Skipping unwanted buffers
Added a function which checks if a buffer is skippable or not and implemented this function into the
next/previous-buffer function. Now there is a simple list of exceptions to *X* buffers and to normal
buffers that should be kept/skipped.

Furthermore, I fixed the documentation strings for all functions, so that emacs is no longer
complaining about them