Added serveral UI improvements and changes
Changes the modeline to doom-modeline which look quite nice. I also added dashboard and
all-the-icons so that the starting buffer is no longer *scratch* but a nice dashboard with my latest
files and stuff like that. I reformatted my theme-config file and added some mode-line configuration
to the custom file.

I also deactivated org-drill for now since it is not being used.